Create search in rotated sorted array
The only difference is that due to the existence of duplicates,
arr[l] == arr[mid] could be possible, the first half could be out of order
(i.e. not in the ascending order, e.g. {3, 1, 2, 3, 3, 3, 3})
we have to deal this case separately.
In that case, it is guaranteed that arr[high] also equal to arr[mid],
so the condition arr[mid] == arr[l] == arr[r] can be checked before the original logic,
if so then move left and right both towards the middle by 1 and repeat.
Closer look to the duplicate case

let go in detail...

{3, 1, 2, 3, 3, 3, 3})

I have taken the same eg as above,

so we know when we divide array,

either of 1 half has to be sorted because since the array is rotated, still one half will be sorted and i fell you understand this case,

now the problem is , as in above case if i some how find two halves let say

3 1 2 3 is one half, and 3 3 3 is another half and our mid is element 3 also it at index '3'

ok i hope you get it

now,

can we distingusih over which half we should minimize are search space, can we ?

the answer is no

because, if i compared it with nums[low] it is 3 and if i look for nums[high] it is again 3

so our search space will not be divide correctly as both are same what is correct half we cannot determine and we cannot get to know which is our sorted half and i feel you are getting it now,

so we have to get rid of the condition nums[low] == nums[mid] == nums[high] seperately.

According to our code, it will run the if condition which will compare with nums[low] because if condition will run first , but again we cannot determine it would have the correct partition.

so that's why we trimmed our search space both from start and end so that we can determine the valid partition and get to determine the sorted half.

and handled the case if nums[low] == nums[mid] == nums[high] seperately.

and the worst time complexity is ~ O(n/2) of this entire code, because of duplicate case. average tc is O(logn)

ALGORITHM

Check for if target == nums[mid], if found return true.
With duplicates we can have nums[l] == nums[r] so just update the l and r towards middle.
If not than check for first half is that in order (i.e nums[l] <= nums[mid]
check if target is in the range of [left, mid -1], if((nums[l] <= target) && (nums[mid] > target)) and search in it.
if not in first half search in second half.
Check if target is in range of [mid + 1, right], if((nums[mid] < target) && (nums[r]>= target)) and search in it.
Time Complexity - >
Average O(log n)
worse ~(O(n) = O(n)

space Complexity -> O (1)